+++ tests/conftest.py
@@ -6,7 +6,6 @@ pytest_plugins = ["pytester"]
 @pytest.fixture
 def item_names_for(testdir):
     def _item_names_for(tests_content):
-        # some strange code to extract sorted items
         items = testdir.getitems(tests_content)
         hook = testdir.config.hook
         hook.pytest_collection_modifyitems(session=items[0].session,